refactor: Move maps page to Maps folder
