+++ lib/assets/javascripts/new-dashboard/pages/Maps/Maps.vue
