+++ lib/assets/javascripts/new-dashboard/router/index.js
@@ -4,7 +4,7 @@ import Router from 'vue-router';
 // Pages
 import Home from 'new-dashboard/pages/Home/Home';
 import Solutions from 'new-dashboard/pages/Solutions';
-import Maps from 'new-dashboard/pages/Maps';
+import Maps from 'new-dashboard/pages/Maps/Maps';
 import Data from 'new-dashboard/pages/Data';
 import Search from 'new-dashboard/pages/Search';
 import Notifications from 'new-dashboard/pages/Notifications';
